Dynamic init containers list (#52)
This PR moves the static list of init containers on the statefulset into
the values file and uses a generic init container helper template so
that more complex and dynamic init container behavior may be achieved.

The upside is highly flexible startup behavior. The down side is that if
you have a non-default use case, you will need to duplicate _every_ init
container config in your values because helm [does not allow amending
lists](https://github.com/helm/helm/issues/6569)

Though it can be managed with careful use of the `--set`
[command](https://stackoverflow.com/questions/65011579/how-replace-specific-property-value-in-an-arrays-item-in-helm-values-yaml-on-co)